Changing references from `i32` to `usize` in spanning methods.

+++ src/tokenize/api.rs
@@ -2,20 +2,20 @@ use tokenize::util::string_span_tokenize;
 
 // TODO: determine if we should implement an iterator
 /*struct Tokens {
-    inner_vec: Vec<i32, i32>,
-    curr_idx: i32,
+    inner_vec: Vec<usize, usize>,
+    curr_idx: usize,
 }
 
 impl Tokens {
-    fn new(inner_vec: Vec<i32, i32>) -> Tokens {
+    fn new(inner_vec: Vec<usize, usize>) -> Tokens {
         Tokens { inner_vec: inner_vec, curr_idx: 0 }
     }
 }
 
 impl Iterator for Tokens {
-    type Item = Vec<i32, i32>;
+    type Item = Vec<usize, usize>;
 
-    fn next(&mut self) -> Option<Vec<i32, i32>> {
+    fn next(&mut self) -> Option<Vec<usize, usize>> {
         if curr_idx >= inner_vec.len() {
             None
         } else {
@@ -29,7 +29,7 @@ impl Iterator for Tokens {
 pub trait TokenizerI {
     fn tokenize<'a>(&'a self, s: &'a str) -> Vec<&str>;
 
-    fn span_tokenize(&self, s: &str) -> Vec<(i32, i32)>;
+    fn span_tokenize(&self, s: &str) -> Vec<(usize, usize)>;
 
     fn tokenize_sents<'a>(&'a self, strings: Vec<&'a str>) -> Vec<Vec<&str>> {
         let mut vec = Vec::new();
@@ -41,10 +41,10 @@ pub trait TokenizerI {
 
     // TODO: Unsure if we can use an iterator the way Python does...
     // so for now, just return a Vec
-    fn span_tokenize_sents(&self, strings: Vec<&str>)  -> Vec<Vec<(i32, i32)>> {
-        let mut result: Vec<Vec<(i32, i32)>> = Vec::new();
+    fn span_tokenize_sents(&self, strings: Vec<&str>)  -> Vec<Vec<(usize, usize)>> {
+        let mut result: Vec<Vec<(usize, usize)>> = Vec::new();
         for s in strings.iter() {
-            let span: Vec<(i32, i32)> = self.span_tokenize(s);
+            let span: Vec<(usize, usize)> = self.span_tokenize(s);
             result.push(span);
         }
         return result;
@@ -60,10 +60,10 @@ impl TokenizerI for StringTokenizer {
         return split_str;
     }
 
-    fn span_tokenize(&self, s: &str) -> Vec<(i32, i32)> {
-        let mut result: Vec<(i32, i32)> = Vec::new();
+    fn span_tokenize(&self, s: &str) -> Vec<(usize, usize)> {
+        let mut result: Vec<(usize, usize)> = Vec::new();
         match string_span_tokenize(s, self._string) {
-            Ok(spans) => { 
+            Ok(spans) => {
                 for span in spans.iter() {
                     result.push(span.to_owned());
                 }
@@ -73,5 +73,3 @@ impl TokenizerI for StringTokenizer {
         return result;
     }
 }
-
-
